pyproject.toml
[project]
name = "animepahe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "httpx>=0.28.1",
    "m3u8>=6.0.0",
    "pycryptodomex>=3.23.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "readchar>=4.2.1",
    "rich>=14.2.0",
]
[project.scripts]
ani = "animepahe.main:start"
